deployement/requirements.txt
absl-py==2.1.0
aioquic==1.0.0
asgiref==3.8.1
astunparse==1.6.3
attrs==23.2.0
beautifulsoup4==4.12.3
blinker==1.8.2
Brotli==1.1.0
certifi==2024.6.2
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
cryptography==42.0.8
ffmpeg-python==0.2.0
filelock==3.15.1
Flask==3.0.3
flatbuffers==24.3.25
fsspec==2024.6.0
future==1.0.0
gast==0.5.4
google-pasta==0.2.0
grpcio==1.64.0
h11==0.14.0
h2==4.1.0
h5py==3.11.0
hpack==4.0.0
hyperframe==6.0.1
idna==3.7
ImageHash==4.3.1
imageio==2.34.1
itsdangerous==2.2.0
Jinja2==3.1.4
kaitaistruct==0.10
keras==3.3.3
lazy_loader==0.4
ldap3==2.9.1
libclang==18.1.1
llvmlite==0.43.0
lxml==5.2.1
m3u8==5.0.0
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
mitmproxy==10.3.1
mitmproxy-macos==0.5.2
mitmproxy_rs==0.5.2
ml-dtypes==0.3.2
more-itertools==10.3.0
mpmath==1.3.0
msgpack==1.0.8
namex==0.0.8
networkx==3.3
numba==0.60.0
numpy==1.26.4
openai-whisper==20231117
opencv-python==4.9.0.80
opencv-python-headless==4.10.0.84
opt-einsum==3.3.0
optree==0.11.0
outcome==1.3.0.post0
packaging==24.0
pandas==2.2.2
passlib==1.7.4
pdf2image==1.17.0
pdf2pptx-cli==0.1.2
pillow==10.3.0
protobuf==4.25.3
publicsuffix2==2.20191221
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
Pygments==2.18.0
pylsqpack==0.3.18
PyMuPDF==1.24.2
PyMuPDFb==1.24.1
pyOpenSSL==24.1.0
pyparsing==3.1.2
PyPDF2==3.0.1
pyperclip==1.8.2
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-pptx==0.6.23
pytz==2024.1
PyWavelets==1.6.0
regex==2024.5.15
reportlab==4.2.0
requests==2.31.0
rich==13.7.1
ruamel.yaml==0.18.6
ruamel.yaml.clib==0.2.8
scikit-image==0.23.2
scipy==1.13.0
selenium==4.21.0
service-identity==24.1.0
setuptools==70.0.0
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.5
sympy==1.12.1
tensorboard==2.16.2
tensorboard-data-server==0.7.2
tensorflow==2.16.1
termcolor==2.4.0
tifffile==2024.4.18
tiktoken==0.7.0
torch==2.3.1
tornado==6.4
tqdm==4.66.4
trio==0.25.1
trio-websocket==0.11.1
typing_extensions==4.12.0
tzdata==2024.1
Unidecode==1.3.8
urllib3==2.2.1
urwid-mitmproxy==2.1.2.1
webdriver-manager==4.0.1
Werkzeug==3.0.3
wget==3.2
wheel==0.43.0
wrapt==1.16.0
wsproto==1.2.0
XlsxWriter==3.2.0
zstandard==0.22.0
